Preparing for refactor.
This refactor is meant to adjust the project structure to take better advantage
of an ES Module buildless browser workflow as well as allow Deno bundling from
its entry points.

The current code base has a lot of duplicate references that collide with each
other upon bundling. The results are... less than ideal.