Rename Spacing formula to TextUnits.
This commit also recategorizes it as a text formula instead of a layout
formula. This is because I use `ex` units for a lot more than spacing.

It's a rather versatile unit. It can be used for spacing, vertical rhythm,
border radii, border widths, and its values are always derived from the
approximate x-height of the current font.

So, from the base, each value in the scale is treated as a *unit* of text.