Refactor extractor and parser to use the *result* of validator.
I realized that in v1, the conversion pipeline was sometimes unpredictable.

This is because internally the `extractor()` and `parser()` were not *reusing* the
result of `validator()`. They were calling it independently. This meant that the
color conversion was actually being performed on *completely different*
validated colors than the user actually input.

This didn't hinder the functionality too much, but it was incredibly apparent
when errors did occur. The sensible refactor here is to treat the color
conversion steps as their own sequential operations that pass a *single*
validated result through extraction, through parsing, to conversion.

If a color isn't of a valid format, it doesn't get passed along to the extractor
*at all*. This makes the old implementation completely uneccesary as color
validation is already the root condition of successful color conversion.