Begin refactoring palette formulas as palette_create() utility.
I realized it didn't make sense to provide separate utilities for generating
variants and material palettes, because they all work to adhere to only two core
palette configurations: artistic and material-esque.

The idea of formulas in general are a poor fit for the library's intended
purpose. They work by defining structure first and slotting data in, when the
way it should actually work is to derive structure FROM the data.

So this commit begins the refactor for that, and by the end, will eliminate the
use case for palette formulas entirely while also making palette generation as a
whole more flexible.

Win-win, don't you think?