Staging the color utilities.
This will give me a better idea of exactly how color data is best generated and
structured while hopefully clarifying how that data needs to render.

This staging deployment is totally unpolished, so I recommend not touching
it. It's out here so I can lab and improve the workflow in a real browser.