Add fn_propagate() functional utility.
This utility performs the same function as $_ and $$_ propgation methods of the
object factory. However, because it does not have to track state and the emitter
is passed in directly, it doesn't have to guess anything about the data it
consumes.

Also, because it's a functional action, it can be preset and converted to an
emitter itself. This makes it eligible for including in a process or pipeline.