Refactor palette_create() for simplicity.
I also reworked the generated palettes to allow for some degree of data
redundancy. I realized it's much less expensive to check if an array is empty or
not than to respond to varying data configurations.

This change makes the output structure more predictable, which greatly reduces
the difficulty of assembling the token structure later.