Refactor color token generation.
The major change is that I've rearranged things so the user is never actually
exposed to the palette data. It wasn't really providing anything useful, so it
acted more as noise.

The refactor allows you pass in a color, set the configuration for your output,
and even check it against various perceptual conditions before filtering for
accessibility.

This also means the color API as been reduced to the following functions:

+ convert(to, color)
+ adjust(settings, color)
+ mix(settings, color)
+ harmony(settings, color)
+ palette(settings, color)
+ output(format, dict)

Basically, I've reduced it down to the conversion function, adjusting and
blending, generating base harmonies, generating the tokens themselves, and
preparing them for export.